// src/types/categories.ts
export enum Categories {
  KIDS = 'KIDS',
  CLOTHES = 'CLOTHES',
  FOOTWEAR = 'FOOTWEAR',
  PLUS_SIZE = 'PLUS_SIZE',
  ACCESSORIES = 'ACCESSORIES',
}

export enum CategoriesTitle {
  KIDS = 'Infantil',
  CLOTHES = 'Roupas',
  FOOTWEAR = 'Calçados',
  PLUS_SIZE = 'Plus size',
  ACCESSORIES = 'Acessórios',
}

export enum Clothes {
  SAIA = 'Saia',
  BODY = 'Body',
  SHORT = 'Short',
  CALCA = 'Calça',
  BLUSA = 'Blusas',
  CROPPED = 'Cropped',
  MACACAO = 'Macacão',
  VESTIDO = 'Vestido',
  BIQUINI = 'Biquíni',
  CONJUNTO = 'Conjunto',
  MACAQUINHO = 'Macaquinho',
}

export enum Accessories {
  BOLSA = 'Bolsa',
  COLAR = 'Colar',
  CINTO = 'Cinto',
  BRINCO = 'Brinco',
}

export enum Footwear {
  TENIS = 'Tênis',
  RASTEIRINHA = 'Rasteirinha',
}

export enum PlusSize {
  SHORT = 'Short',
  VESTIDO = 'Vestido',
  BLUSA = 'Blusa',
  MACAQUINHO = 'Macaquinho',
}

export enum AllItemsByCategories {
  SAIA = 'SAIA',
  BODY = 'BODY',
  CALCA = 'CALCA',
  BOLSA = 'BOLSA',
  COLAR = 'COLAR',
  BLUSA = 'BLUSA',
  CINTO = 'CINTO',
  TENIS = 'TENIS',
  SHORT = 'SHORT',
  BRINCO = 'BRINCO',
  VESTIDO = 'VESTIDO',
  CROPPED = 'CROPPED',
  MACACAO = 'MACACAO',
  BIQUINI = 'BIQUINI',
  CONJUNTO = 'CONJUNTO',
  MACAQUINHO = 'MACAQUINHO',
  RASTEIRINHA = 'RASTEIRINHA',
}

export enum AllItemsByCategoriesTitle {
  SAIA = 'Saia',
  BODY = 'Body',
  CALCA = 'Calça',
  BOLSA = 'Bolsa',
  COLAR = 'Colar',
  BLUSA = 'Blusa',
  CINTO = 'Cinto',
  TENIS = 'Tênis',
  SHORT = 'Short',
  BRINCO = 'Brinco',
  VESTIDO = 'Vestido',
  CROPPED = 'Cropped',
  MACACAO = 'Macacão',
  BIQUINI = 'Biquíni',
  CONJUNTO = 'Conjunto',
  MACAQUINHO = 'Macaquinho',
  RASTEIRINHA = 'Rasteirinha',
}